pyproject.toml
[project]
name = "csv-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandera>=0.27.1",
    "polars>=1.36.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.2",
]
